trying to delete newly created and empty records, but it is not working; current problem is that created==True only the first time the view is called, but not the second time when the form is submitted

+++ tracker/views.py
@@ -84,14 +84,16 @@ def record_data(request, pk, year, month, day):
     habit = get_object_or_404(Habit, pk=pk)
     record_date = date(year, month, day)
     record, created = DailyRecord.objects.get_or_create(habit_id=habit, date=record_date)
+    breakpoint()
     if request.method == "GET":
         form = DailyRecordForm(instance=record)
     else:
         form = DailyRecordForm(data=request.POST, instance=record)
-        if form.is_valid():
-            form.save()
-            return redirect('habit_details', pk=pk)
-        elif created:
+        breakpoint()
+        if form['quantity'].value() == '' and created:
             record.delete()
+            return redirect('habit_details', pk=pk)
+        elif form.is_valid():
+            form.save()
             return redirect('habit_details', pk=pk)
     return render(request, 'tracker/record_data.html', {"form": form, "habit": habit, "record_date": record_date, "record": record, "created":created})
